+++ packages/grid/src/components/cell-editors/select/select-editor.component.ts
@@ -10,7 +10,7 @@ import { AITableSelectField } from '../../../types';
 import { SelectOptionComponent } from '../../cell-views/select/option.component';
 import { AbstractEditCellEditor } from '../abstract-cell-editor.component';
 import { ThyFormModule } from 'ngx-tethys/form';
-import { AITableQueries } from '../../../utils';
+import { AITableQueries } from '../../../core';
 @Component({
     selector: 'select-cell-editor',
     templateUrl: './select-editor.component.html',