+++ packages/state/src/action/field.ts
@@ -1,8 +1,9 @@
 import { AITableQueries } from '@ai-table/grid';
-import { ActionName, AddFieldAction, RemoveFieldAction, SetFieldAction, AIViewTable } from '../types';
+import { ActionName, AddFieldAction, RemoveFieldAction, SetFieldAction } from '@ai-table/utils';
 import { AITableViewField, AITableField, IdPath, NumberPath } from '@ai-table/utils';
 import { isPathEqual } from '../utils';
 import { getFieldPositionInView } from '../utils/field/position-field';
+import { AIViewTable } from '../types/ai-table';
 
 export function addField(aiTable: AIViewTable, field: AITableField, originId?: string, isCopy?: boolean) {
     const operation: AddFieldAction = {