+++ packages/state/src/constants/index.ts
@@ -1,4 +1,3 @@
-export * from './default';
 export * from './field-menu-item';
 export * from './context-menu-item';
 export * from './view';