+++ src/app/service/table.service.ts
@@ -48,6 +48,8 @@ export class TableService {
 
     rowDragDisabled: WritableSignal<boolean> = signal(false);
 
+    hiddenIndexColumn: WritableSignal<boolean> = signal(false);
+
     records!: WritableSignal<AITableViewRecords>;
 
     fields!: WritableSignal<AITableViewFields>;
@@ -121,6 +123,10 @@ export class TableService {
 
     setRowDragDisabled(rowDragDisabled: boolean) {
         this.rowDragDisabled.set(rowDragDisabled);
+    }
+
+    setHiddenIndexColumn(hiddenIndexColumn: boolean) {
+        this.hiddenIndexColumn.set(hiddenIndexColumn);
     }
 
     setActiveView(activeViewId: string) {